extract ConversationMembers
