+++ wire-ios-request-strategy/Sources/Payloads/Payload+Conversation.swift
@@ -189,82 +189,6 @@ extension Payload {
         }
     }
 
-    struct Service: Codable {
-        let id: UUID
-        let provider: UUID
-    }
-
-    struct ConversationMember: CodableEventData {
-
-        enum CodingKeys: String, CodingKey {
-            case id
-            case qualifiedID = "qualified_id"
-            case target
-            case qualifiedTarget = "qualified_target"
-            case service
-            case mutedStatus = "otr_muted_status"
-            case mutedReference = "otr_muted_ref"
-            case archived = "otr_archived"
-            case archivedReference = "otr_archived_ref"
-            case hidden = "otr_hidden"
-            case hiddenReference = "otr_hidden_ref"
-            case conversationRole = "conversation_role"
-        }
-
-        static var eventType: ZMUpdateEventType {
-            return .conversationMemberUpdate
-        }
-
-        let id: UUID?
-        let qualifiedID: QualifiedID?
-        let target: UUID?
-        let qualifiedTarget: QualifiedID?
-        let service: Service?
-        let mutedStatus: Int?
-        let mutedReference: Date?
-        let archived: Bool?
-        let archivedReference: Date?
-        let hidden: Bool?
-        let hiddenReference: String?
-        let conversationRole: String?
-
-        init(id: UUID? = nil,
-             qualifiedID: QualifiedID? = nil,
-             target: UUID? = nil,
-             qualifiedTarget: QualifiedID? = nil,
-             service: Service? = nil,
-             mutedStatus: Int? = nil,
-             mutedReference: Date? = nil,
-             archived: Bool? = nil,
-             archivedReference: Date? = nil,
-             hidden: Bool? = nil,
-             hiddenReference: String? = nil,
-             conversationRole: String? = nil) {
-            self.id = id
-            self.qualifiedID = qualifiedID
-            self.target = target
-            self.qualifiedTarget = qualifiedTarget
-            self.service = service
-            self.mutedStatus  = mutedStatus
-            self.mutedReference = mutedReference
-            self.archived = archived
-            self.archivedReference = archivedReference
-            self.hidden = hidden
-            self.hiddenReference = hiddenReference
-            self.conversationRole = conversationRole
-        }
-    }
-
-    struct ConversationMembers: Codable {
-        enum CodingKeys: String, CodingKey {
-            case selfMember = "self"
-            case others
-        }
-
-        let selfMember: ConversationMember
-        let others: [ConversationMember]
-    }
-
     struct ConversationTeamInfo: Codable, Equatable {
         enum CodingKeys: String, CodingKey {
             case teamID = "teamid"
